requirements.txt
# ============================================================
# requirements.txt — Mental Health Chatbot
# Python >= 3.11
# ============================================================

# ── Web Framework ───────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.29.0

# ── Auth ────────────────────────────────────────────────────
python-jose[cryptography]
passlib[bcrypt]

# ── LLM ─────────────────────────────────────────────────────
openai>=1.0.0

# ── Embedding & Emotion ─────────────────────────────────────
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.0

# ── Vector Search ───────────────────────────────────────────
faiss-cpu>=1.7.4
qdrant-client>=1.8.0

# ── Database ─────────────────────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
aiosqlite>=0.20.0

# ── PDF ──────────────────────────────────────────────────────
pypdf>=4.0.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.26.0
httpx>=0.27.0
requests>=2.31.0